Add files via upload
The Twitter Application Programming Interface (API) allows users, businesses, and developers to have automated access to the public data that Twitter users have given consent to disclose with the rest of the world. This information includes tweets, user profiles, and other publicly available information.
TWITTER DEVELOPER ACCOUNT:

It is necessary to have a Twitter developer account to collect tweets from other users. To apply for a developer account on Twitter, connect to developer.twitter.com, while logged into your existing Twitter account. Follow the on-screen instructions to complete the application. After clicking the sign-up button, enter the name of the developer account, the location, and the details of the use case. Check the developer agreement, make sure it's acceptable, and then submit it. After signing up for a Twitter developer account, which is completely free of charge and grants rapid access to the Twitter API, we are granted "essential access."

ESSENTIAL ACCESS:
There are now two additional access levels that each provide you with more data at no additional cost.
Essential access grants the user immediate access to the Twitter API version 2, one App environment, and the potential to retrieve up to 500k Tweets each month. This ought to satisfy the requirements of most developers, particularly in the beginning stages. 

ELEVATED ACCESS:
So, we need to apply for elevated access. Explain why you need an elevated access for your use case properly mentioning that you’re a student and the clear intention is to research for a thesis project. This access including the capability to retrieve up to two million Tweets every single month and three different App environments (development, staging, and production). If you are already using Twitter API version 2, you will see an upgrade to Elevated access for your Projects today. This will take place automatically. In that case, you will be required to apply for Elevated access. Twitter is working on a way to enable developers who have previously had their requests for a developer account denied applying for an account again. Soon, it will provide a mechanism that enables reapplication for Elevated access or signup for Essential access. Twitter API has different credentials compared to other API’s. API keys are used to build the apps and projects in developer portal. 

CONSUMER KEYS:
There are four types of secret keys and tokens which are unique to every user. Think of these as the username and password that represents the app when making API requests.

AUTHENTICATION TOKENS:
BEARER TOKEN: Bearer token authenticates requests on behalf of the developer app.

ACCESS TOKENS AND SECRET KEYS:
To authenticate OAuth 1.0a API, users must have an Access Token and a Secret. They name the Twitter account on whose behalf the request is made.

CHOICE OF CRYPTOCURRENCY: Bitcoin
Cryptocurrencies such as Bitcoin (BTC) attracted a lot of attention in recent months due to their unprecedented price fluctuations.

CHOICE OF INFLUENCERS: Influencers selection has been made based on the number of followers.
1.	Anthony Pompliano.                       
2.	Andreas M. Antonopoulos
3.	Roger Keith Ver
4.	Barry Silbert
5.	Vitalik Buterin
6.	Michael J. Saylor
7.	Bobby Lee
8.	Charlie Lee
9.	Jack Dorsey
10.	Bitcoin Archive
11.	Tyler Winklevoss
12.	Cameron Winklevoss
13.	Scott Melker
14.	Lex Moskovski